Change Mem::cstr_at_utf8() to return a Result rather than panicking
This function is used a lot in debug logging to make strings readable,
but these strings aren't always required to be UTF-8, so in some rare
case we might have a panic due to that. Returning an `Err()` with the
byte slice makes this kind of use safer.

Moving the `.unwrap()` to the caller also makes it more obvious a panic
could happen, and hopefully discourages unnecessarily requiring UTF-8.