Use glCopyTexImage2D() instead of glBlitFramebufferEXT()
The latter doesn't seem to be available in OpenGL ES 1.1, but the former
is, so this gets us one step closer to making our presentRenderbuffer:
implementation work on true OpenGL ES (e.g. on Android).

Thanks to @ciciplusplus for pioneering this approach when porting to
Android. I wouldn't have had the idea otherwise.