Split touchHLE crate into lib and bin targets
This is in preparation for Android support. Android apps can't be
native code executables, but they can load a native code library, so we
need a library target on Android. Alas, we we can't make the crate type
(executable or library) conditional on the platform, so we have to have
it elsewhere too. Doing this split early makes for a clearer history.

Until now, there was only an executable target, so rustdoc included
private items by default. Now that there's a library target, that
default doesn't apply, which means we would get almost no documentation!
The .cargo/config.toml file is added to override this, though this
now means we get private items in _other_ crates documented too, and
we have to suppress warnings about links to private items that didn't
happen before.

src/main.rs is renamed to src/lib.rs because almost all the code is now
in the library. src/bin.rs is the new module for the executable. It
could have been named src/main.rs, but then git wouldn't realise
src/lib.rs is the successor of the old src/main.rs, which would be
annoying for blames and merges.