Introduce macro for non-warning/error output, use it everywhere
This is in preparation for Android support. Printing to stdout/stderr
seems to do nothing on that platform, so we'll need to output things in
a different way there.

This also definitively answers the question of whether to use stdout or
stderr (stderr, always) and means we could easily make future
enhancements like prefixing everything with touchHLE or redirecting to a
file.