Implement Android support
This is a squashed merge of the Android branch, the final commit of
which was 3c517a36b1710f771f38e783afca6417624d26f8, tagged as
android-before-merge.

@ciciplusplus took the initiative to start porting touchHLE to Android
and got the first version working. @hikari-no-yume's contributions came
later and are mostly fixes and cleanups of various kinds.

There are still various unfinished things in this port, but it's now
in a clean enough state that it no longer needs its own branch.

(Merged pull request by @ciciplusplus, resolves #88.)

Co-authored-by: ciciplusplus <ciciplusplus@gmail.com>