Render splash screen using OpenGL ES
Previously OpenGL 3.2 was used. The hope was that on Android, a similar
OpenGL *ES* version could be used, but this apparently isn't practical
because we seem to have to use the same OpenGL ES version for all
contexts in our window, so it would prevent the app from using
OpenGL ES 1.1. :(

So instead let's use OpenGL ES 1.1, via our GLES1-on-GL2 layer where
appropriate, so that we can display the splash screen on all platforms.
This also allows us to share more code between displaying splash screens
and presenting renderbuffers.

With this commit, the current division of OpenGL-related things between
touchHLE::window and touchHLE::frameworks::opengles no longer makes much
sense. They will be reorganized in a later commit that doesn't have
functional changes.