Implement compositor for Core Animation layers
Until now, touchHLE not had any general support for displaying Core
Animation layers; the only supported case was apps that draw everything
using OpenGL ES and present it to an CAEAGLLayer that covers the entire
screen. That pattern is common in games, but it excludes any app that
uses UIKit for its UI, since UIKit uses one layer per view.

This commit changes this by introducing a compositor. If there isn't
an opaque CAEAGLLayer that covers the entire screen, Core Animation
will take over repainting the window and try to composite the various
layers (some or all of which might not be EAGL layers) at 60fps. On the
other hand, if there is such a layer, touchHLE continues to present it
directly to the window for the sake of efficiency.

The compositor supports all the properties currently exposed on CALayer:
layers may have a background color, may display OpenGL ES content, and
basic positioning, scaling, opacity and opaqueness work.

Note that while empty UIKit views can be displayed now, they still can't
be interacted with.

Also, unfortunately, compositing OpenGL ES content relies on
glReadPixels(), which is extremely slow. :(