Implement a UIKit-based app picker, replacing the old message box one
This new app picker is prettier, more user-friendly, consistent across
platforms, makes use of apps' display names and icons, and can display
more apps at once (though there's still a hard limit, for now). It also
provides a base on which to build additional UI features. One of these
is added in this commit: a link to the website.

The new app picker is more or less its own app, but since it is entirely
host code, it was necessary to add a way to create an Environment
without loading a guest binary. This is done in a fairly generic way and
could be reused for unit tests in future.

Closes #28.