Fix initializer chains in UIView (do not call init from initWithFrame:)
3b606c7fbed1f545d3976b2823e8b0b056fe3c7f accidentally caused a
regression in Mystery Mania, because that game has a view that
super-calls initWithCoder:, which then called init, which was
overridden by the class.

This commit changes UIView and all its subclasses to be more careful
in their constructors. Constructors should never make ordinary calls,
only supercalls, and for UIView, the only valid supercalls should be
initWithCoder: and initWithFrame:.