Add Mystery Mania to the supported games and changelog
The credit here goes to @KiritoDv because they were the one who decided
to try to get this game working, and wrote a lot of patches that got it
most of the way there. Some of them were merged by me, and some were
replaced by my own patches. My main contribution was getting the audio
to work. I don't want to steal the spotlight. :)