Fix division by zero calculating fog
Check if both start and end distances are equal. If so, set the start
distance to be slightly less than the end distance to prevent
a division by zero thats handled differently in different
implementations.
This fixes an issue in Doom 2 RPG where levels would be pitch black
when both start and end distances for the fog were set to 10000.