Add license text viewer
This is needed on Android, where the command-line flag for viewing
license text isn't actually accessible.

This commit also updates the license code text to include the dylib and
font licenses on Android, since they're bundled in the APK and therefore
wouldn't otherwise be user-accessible on that platform.

The license code text is also slightly adjusted to be prettier when
viewed in the new viewer.