Add runtime type checking for host-to-host messages in debug builds
In debug builds, calls to msg_send are replaced with msg_send_debug,
which adds the (hashed) typeid of the arguments. HostIMPs can also
generate the typeid of their arguments, and can crash on debug builds.
The msg_unchecked! macro (and friends) opt out of this behaviour for the few cases
where we want to avoid this behaviour, currently only in static string
allocation and dictionaryWithObjectsAndKeys:.