Move the code-of-conduct to its own file, add more trans rights 🏳️‍⚧️
I'm going to add more stuff to CONTRIBUTING.md soon, so this helps to
avoid it getting too crowded. Also, some people seem to use the
existence of a CODE_OF_CONDUCT.md as a simple heuristic for whether it's
safe to contribute, and I want to make it absolutely clear that I am on
their side there.

Change-Id: Id83ca9cae5d89379577143d0460ed1bf02da56f3