Implement AudioQueueGetProperty() and AudioQueueGetPropertySize()
Since 48a5dd4d297b1c34909ea7f4ac19ad3b9458e0f1 added support for audio
queue property listeners for kAudioQueueProperty_IsRunning, Touch & Go
has been broken because it calls AudioQueueGetProperty() for that same
property in its property listener callback. This commit fixes this.
It doesn't use AudioQueueGetPropertySize() but it seemed like a good
idea to add it too.

The code is modelled on AudioFileGetProperty() and
AudioFileGetPropertyInfo(). For some reason the naming and parameter
ordering is not fully consistent between the APIs?

Change-Id: I5d9a3f189bff6174fdd955d3c77411d9a086e080