Implement arbitrary affine transforms for CGContext and CGBitmapContext
This necessitated a radical change to using a per-glyph rather than
per-pixel callback for Font::draw() to enable sampling the glyphs with
an arbitrary scale or orientation. In the process, part of the clipping
responsibility was moved to the caller, and vertical clipping has been
added (previously there was only horizontal clipping).

Change-Id: Ibe81dad79d84b43dca2d37435cd75346413cc0c1