Include trailing and leading whitespace in text layout metrics
RustType's PositionedGlyph::pixel_bounding_box() was used for
calculating the width of lines of text. This worked fine in most cases,
but because it returns None for whitespace characters, leading and
trailing whitespace was effectively ignored when measuring a line's
width. This is fixed by using ScaledGlyph::h_metrics() instead, which
doesn't exclude whitespace characters.

This fixes the issue in I Love Katamari where spaces were displayed as
zero-width, because that game generates a proportional sprite font from
the system font by rendering each character individually.

This commit also changes the handling of whitespace for word wrapping:
whitespace needs to be trimmed at word wrap points, because it will no
longer be ignored by the width measurement.

Some care has been taken to try to avoid noticeably changing the layout
of text that lacks leading/trailing whitespace, but there are some small
changes here and there.

Change-Id: I113b801754f686a388413134bb1479469d95de52