Update dylibs to iPhone OS 3.0.1 and claim iPhone OS 3.0 support
This is the last iPhone OS version (as far as I know) that has actual
dylibs in its OS image that we can use. iPhone OS 3.1 added the dyld
shared cache, which means there are no longer individual .dylib files,
but rather a massive über-dylib that combines all the libraries into
one, including all the non-GPL ones, so it'd be a problem from a
copyright perspective. Hence sticking to iPhone OS 3.0.1 for now.

The choice of the iPhone 3G this time is because iPhone OS 3 was a paid
upgrade for iPod touch owners, so it's harder to get hold of a system
image for the iPod touch than for the iPhone.

Future work: maybe replace this with the iPhone 3GS version so we also
have ARMv7 slices; I'm not so comfortable with mixing ARMv6 dylibs with
ARMv7 app code.

Change-Id: Ib115c7430c8fa79c6192c5ae0144c478eefbf577