Remove '(git)' suffix from git versions, add suffix to non-git versions
This change is in preparation for using GitHub Actions for release
builds, and also reflects the fact that git is more likely to be
correct about the version number than Cargo.toml.

Since this creates a risk that releases will be tagged without updating
the version number in Cargo.toml (release builds had the .git directory
removed previously), a warning has been added for when the numbers do
not match.

Change-Id: I9d29b3d57a41477e1bd954ac3df84a665ae8126c