Adjust build.rs for wrappers to use correct target os
Since build.rs is built for the host machine. cfg target os refers to
the host running the build. We want the lib names to be adjusted based
off of the compilation target instead.

This fixes some of the compilation issues when building for android on
windows.

Co-authored by: ciciplusplus <ciciplusplus@gmail.com>

Change-Id: Ib9fa91c93710399b6b40d8472c4a1c98ad44bd3c