Add nix flake and direnv
The motivation behind this change is to ease setting up a development
environment for people on Linux or MacOS while also providing a package
for Nix/NixOS users.

This is done using a nix flake file, which provides build instructions
for people using the nix package manager.

This flake file offers both the development environment and also the
appropriate nix package which can be installed by users of nix by simply
issuing the command (if this were to be merged):

nix profile install git+https://github.com/touchHLE/touchHLE?submodules=1

(The submodules thing is required because nix by default ignores
submodules and this project requires them in order to build)

I think that this is desireable due to the fact that a recent CMake
change broke this project on Arch Linux, this ensures that people using a
distro such as Arch can benefit from stable (vendored) dependencies offered
by nixpkgs. Nix does this by basically tracking it's packages
and dependencies with a git hash as shown in the flake.lock file.

Direnv is just a quality of life tool that automatically integrates said
stable development environment into common tools such as shells, IDEs and
text editors. This offers ease of use to people already familiar with
these tools.

Notes:

 - Currently cargo tests fail while building on nix so I've disabled them,
not sure if that's a problem or not.

 - And the static binary fails to launch the GUI, so I've made this flake
build the dynamic version of the binary instead.

Change-Id: Id4d966a4d90df97d4b556c162187a8c559f793b9