Parse LC_DYLD_INFO to improve 3.1+ compat
Since iPhoneOS 3.1 the app binaries store binding data in the DYLD_INFO command, and leave the external relocations empty in the DYSYMTAB. The mach_object rust crate provides handy parsers for this data.
Currently we only handle the bind data, as rebasing needs slide support and weak and lazy binds are already handled.

Change-Id: I59a65b344eef3942a316f0e16f2263e5ca896d87