Handle entry_size=4 on the symbol stub section
Since iPhoneOS 3.1, the binaries can use a different format for the symbol stub section. This format stores stub offsets in single 4-byte instructions, leaving us with no space for embuing a RET instruction after the SVC, so instead we manually invoke a return after handling the SVC if we marked it with the ret flag.

Change-Id: Ie82030b51bad9520fa4ed16b8a966cbf2a2c6b73